Replicate profiles data.frame

Replicate profiles data.frame

Usage

replicate_profiles(
  profiles,
  replications,
  probabilities = NULL,
  replacement = TRUE,
  seed_used = NULL
)

Arguments

profiles

data.frame of profiles

replications

integer, final number of observations

probabilities

vector of probabilities with the same length as the number of rows of profiles. Does not need to add up to 1 (are reweighted)

replacement

Boolean whether replacement is used

seed_used

Integer with the seed to be used for consistent results

Value

Resampled data.frame of profiles

Examples

replicate_profiles(profiles=data.frame(id=1:100,age=rnorm(100,60,5)),
replications=200,probabilities=rep(1,100))
#>      id      age
#> 1    46 61.15940
#> 2    62 60.20784
#> 3    44 55.36185
#> 4    35 52.47812
#> 5    32 54.16872
#> 6    56 62.22935
#> 7    71 58.86834
#> 8    10 59.38478
#> 9    41 64.38363
#> 10   90 65.62769
#> 11   32 54.16872
#> 12   31 60.04928
#> 13   99 64.22338
#> 14   42 53.74557
#> 15   66 55.57396
#> 16   89 59.34958
#> 17   73 63.61761
#> 18   67 55.00139
#> 19   72 60.25185
#> 20   17 59.61917
#> 21   21 57.35117
#> 22   72 60.25185
#> 23   67 55.00139
#> 24   35 52.47812
#> 25   85 57.05314
#> 26   53 55.91199
#> 27   37 56.55427
#> 28   11 67.65309
#> 29    5 64.38329
#> 30    4 59.64444
#> 31   60 63.63745
#> 32   95 53.96155
#> 33   78 60.49891
#> 34   23 56.65479
#> 35   49 63.46356
#> 36   32 54.16872
#> 37   88 65.02318
#> 38   31 60.04928
#> 39   35 52.47812
#> 40   61 56.62033
#> 41   88 65.02318
#> 42   71 58.86834
#> 43   96 52.30443
#> 44   59 58.70503
#> 45    9 55.82205
#> 46   35 52.47812
#> 47   74 57.14677
#> 48   46 61.15940
#> 49   75 59.23780
#> 50   32 54.16872
#> 51   86 58.83196
#> 52   68 58.07620
#> 53   24 58.86033
#> 54   57 58.28768
#> 55   96 52.30443
#> 56   56 62.22935
#> 57   20 60.78419
#> 58   39 61.01109
#> 59    8 62.01541
#> 60   35 52.47812
#> 61   48 67.81668
#> 62   79 54.22816
#> 63   23 56.65479
#> 64   77 68.81030
#> 65   71 58.86834
#> 66   37 56.55427
#> 67   30 56.28626
#> 68   56 62.22935
#> 69    1 66.81324
#> 70   38 64.38793
#> 71   86 58.83196
#> 72    5 64.38329
#> 73   53 55.91199
#> 74   14 61.10829
#> 75   95 53.96155
#> 76   37 56.55427
#> 77   18 66.51740
#> 78   59 58.70503
#> 79   68 58.07620
#> 80   77 68.81030
#> 81   36 55.55722
#> 82   15 66.51568
#> 83    8 62.01541
#> 84   93 60.05276
#> 85    2 60.87131
#> 86   72 60.25185
#> 87   29 58.12372
#> 88    1 66.81324
#> 89   36 55.55722
#> 90   83 64.00978
#> 91   95 53.96155
#> 92   27 66.04716
#> 93   39 61.01109
#> 94   11 67.65309
#> 95   73 63.61761
#> 96   90 65.62769
#> 97   60 63.63745
#> 98   80 67.61495
#> 99   13 48.87030
#> 100  54 66.94537
#> 101   4 59.64444
#> 102  82 59.51481
#> 103  49 63.46356
#> 104  98 65.86225
#> 105  30 56.28626
#> 106  19 61.63775
#> 107  68 58.07620
#> 108  61 56.62033
#> 109  44 55.36185
#> 110  44 55.36185
#> 111  75 59.23780
#> 112  49 63.46356
#> 113  59 58.70503
#> 114  96 52.30443
#> 115  22 61.16275
#> 116  22 61.16275
#> 117  59 58.70503
#> 118  75 59.23780
#> 119   8 62.01541
#> 120  59 58.70503
#> 121  62 60.20784
#> 122  56 62.22935
#> 123  46 61.15940
#> 124  54 66.94537
#> 125  43 56.87092
#> 126  42 53.74557
#> 127 100 62.20839
#> 128  71 58.86834
#> 129  39 61.01109
#> 130  68 58.07620
#> 131  85 57.05314
#> 132  54 66.94537
#> 133  18 66.51740
#> 134   3 63.05344
#> 135  59 58.70503
#> 136  41 64.38363
#> 137  93 60.05276
#> 138  10 59.38478
#> 139  71 58.86834
#> 140  37 56.55427
#> 141   1 66.81324
#> 142  73 63.61761
#> 143  17 59.61917
#> 144  84 54.21708
#> 145  57 58.28768
#> 146  83 64.00978
#> 147  71 58.86834
#> 148  44 55.36185
#> 149 100 62.20839
#> 150  27 66.04716
#> 151  75 59.23780
#> 152  35 52.47812
#> 153  21 57.35117
#> 154  75 59.23780
#> 155  87 60.55910
#> 156  61 56.62033
#> 157  42 53.74557
#> 158  20 60.78419
#> 159  23 56.65479
#> 160  58 54.71549
#> 161   9 55.82205
#> 162  89 59.34958
#> 163  49 63.46356
#> 164  22 61.16275
#> 165  10 59.38478
#> 166 100 62.20839
#> 167  56 62.22935
#> 168  15 66.51568
#> 169  56 62.22935
#> 170  68 58.07620
#> 171  51 58.24616
#> 172  40 62.92354
#> 173  75 59.23780
#> 174  97 64.28623
#> 175  61 56.62033
#> 176  65 60.79383
#> 177  29 58.12372
#> 178  15 66.51568
#> 179  79 54.22816
#> 180   8 62.01541
#> 181  74 57.14677
#> 182   4 59.64444
#> 183  31 60.04928
#> 184  86 58.83196
#> 185  47 51.04243
#> 186  27 66.04716
#> 187  34 64.79074
#> 188  26 62.81864
#> 189  47 51.04243
#> 190  84 54.21708
#> 191  15 66.51568
#> 192  44 55.36185
#> 193  73 63.61761
#> 194  95 53.96155
#> 195  89 59.34958
#> 196  84 54.21708
#> 197  98 65.86225
#> 198  63 56.52443
#> 199  24 58.86033
#> 200  51 58.24616