Draw from a dirichlet distribution based on mean transition probabilities and standard errors

Draw from a dirichlet distribution based on mean transition probabilities and standard errors

Usage

rdirichlet_prob(n = 1, alpha, se, seed = NULL)

Arguments

n

Number of draws (must be >= 1). If n>1, it will return a list of matrices.

alpha

A matrix of transition probabilities

se

A matrix of standard errors

seed

An integer which will be used to set the seed for this draw.

Value

A transition matrix. If n>1, it will return a list of matrices.

Examples

rdirichlet_prob(n=1,alpha= matrix(c(0.7,0.3,0,0.1,0.7,0.2,0.1,0.2,0.7),3,3),
se=matrix(c(0.7,0.3,0,0.1,0.7,0.2,0.1,0.2,0.7)/10,3,3))
#>           [,1]      [,2]      [,3]
#> [1,] 0.7377871 0.1318882 0.1303247
#> [2,] 0.2129046 0.6058907 0.1812047
#> [3,] 0.0000000 0.2429072 0.7570928

rdirichlet_prob(n=2,alpha= matrix(c(0.7,0.3,0,0.1,0.7,0.2,0.1,0.2,0.7),3,3),
se=matrix(c(0.7,0.3,0,0.1,0.7,0.2,0.1,0.2,0.7)/10,3,3))
#> [[1]]
#>           [,1]      [,2]       [,3]
#> [1,] 0.7947085 0.1056817 0.09960972
#> [2,] 0.2265447 0.5893753 0.18407994
#> [3,] 0.0000000 0.2107516 0.78924840
#> 
#> [[2]]
#>           [,1]      [,2]      [,3]
#> [1,] 0.7787780 0.1037003 0.1175217
#> [2,] 0.2971016 0.5456292 0.1572692
#> [3,] 0.0000000 0.2148933 0.7851067
#>