Replicate profiles data.frame

Replicate profiles data.frame

Usage

replicate_profiles(
  profiles,
  replications,
  probabilities = NULL,
  replacement = TRUE,
  seed_used = NULL
)

Arguments

profiles

data.frame of profiles

replications

integer, final number of observations

probabilities

vector of probabilities with the same length as the number of rows of profiles. Does not need to add up to 1 (are reweighted)

replacement

Boolean whether replacement is used

seed_used

Integer with the seed to be used for consistent results

Value

Resampled data.frame of profiles

Examples

replicate_profiles(profiles=data.frame(id=1:100,age=rnorm(100,60,5)),
replications=200,probabilities=rep(1,100))
#>      id      age
#> 1    81 62.26069
#> 2    15 60.44821
#> 3    61 64.31948
#> 4    37 53.68152
#> 5    69 64.19335
#> 6    38 62.03154
#> 7    82 61.84000
#> 8   100 44.91034
#> 9    50 57.83179
#> 10    9 61.50527
#> 11   15 60.44821
#> 12   91 62.56769
#> 13   73 61.03300
#> 14   89 56.69004
#> 15   92 55.43344
#> 16   49 60.80164
#> 17    8 63.75083
#> 18   73 61.03300
#> 19   29 65.22638
#> 20   34 60.26786
#> 21   40 65.24229
#> 22   35 57.82551
#> 23   11 52.37253
#> 24   60 51.88512
#> 25   77 61.97612
#> 26   53 65.13502
#> 27   29 65.22638
#> 28   46 62.99256
#> 29   91 62.56769
#> 30   94 64.01466
#> 31   54 58.75759
#> 32   78 55.53916
#> 33   78 55.53916
#> 34   85 62.87178
#> 35   96 50.35937
#> 36   44 59.93585
#> 37   62 57.44199
#> 38   24 56.64292
#> 39   87 65.86121
#> 40   94 64.01466
#> 41   60 51.88512
#> 42   67 61.36688
#> 43    8 63.75083
#> 44   26 55.34347
#> 45   52 49.14877
#> 46   53 65.13502
#> 47   78 55.53916
#> 48   28 58.90200
#> 49   70 56.72692
#> 50   57 64.17784
#> 51   45 60.75974
#> 52   88 66.96351
#> 53   13 52.10234
#> 54   33 64.81491
#> 55   76 56.86713
#> 56   56 64.93827
#> 57   19 49.96088
#> 58   70 56.72692
#> 59   12 64.55359
#> 60   30 69.38665
#> 61   74 65.00557
#> 62    6 59.21605
#> 63   12 64.55359
#> 64   43 59.98832
#> 65   61 64.31948
#> 66   21 59.63312
#> 67   52 49.14877
#> 68   84 62.32756
#> 69   11 52.37253
#> 70   49 60.80164
#> 71   90 56.11315
#> 72   45 60.75974
#> 73   29 65.22638
#> 74    7 64.38892
#> 75   90 56.11315
#> 76   16 64.83631
#> 77   33 64.81491
#> 78   70 56.72692
#> 79   30 69.38665
#> 80   44 59.93585
#> 81   50 57.83179
#> 82   26 55.34347
#> 83   38 62.03154
#> 84    8 63.75083
#> 85   38 62.03154
#> 86  100 44.91034
#> 87   21 59.63312
#> 88   89 56.69004
#> 89   13 52.10234
#> 90   21 59.63312
#> 91   75 63.73726
#> 92   29 65.22638
#> 93   23 58.44466
#> 94   60 51.88512
#> 95   10 67.46406
#> 96   49 60.80164
#> 97   32 59.59665
#> 98   91 62.56769
#> 99   23 58.44466
#> 100  54 58.75759
#> 101   3 66.88040
#> 102  54 58.75759
#> 103  40 65.24229
#> 104   2 60.40772
#> 105  58 56.69739
#> 106  46 62.99256
#> 107  63 50.41317
#> 108   7 64.38892
#> 109  67 61.36688
#> 110  34 60.26786
#> 111  17 60.39406
#> 112  23 58.44466
#> 113  39 52.70173
#> 114  69 64.19335
#> 115  17 60.39406
#> 116  42 59.03215
#> 117  28 58.90200
#> 118  36 51.31351
#> 119  13 52.10234
#> 120  31 60.01303
#> 121  15 60.44821
#> 122  47 59.36894
#> 123  75 63.73726
#> 124   7 64.38892
#> 125  41 53.26785
#> 126  86 58.84650
#> 127  41 53.26785
#> 128  47 59.36894
#> 129  97 63.32195
#> 130  79 63.15409
#> 131  49 60.80164
#> 132 100 44.91034
#> 133  31 60.01303
#> 134  20 62.70487
#> 135  79 63.15409
#> 136  89 56.69004
#> 137  33 64.81491
#> 138  93 57.75288
#> 139  93 57.75288
#> 140  82 61.84000
#> 141  37 53.68152
#> 142  69 64.19335
#> 143  28 58.90200
#> 144  97 63.32195
#> 145  14 62.93858
#> 146  34 60.26786
#> 147  64 50.67093
#> 148  26 55.34347
#> 149  94 64.01466
#> 150  61 64.31948
#> 151  38 62.03154
#> 152  62 57.44199
#> 153  62 57.44199
#> 154  90 56.11315
#> 155  65 61.22590
#> 156  88 66.96351
#> 157  94 64.01466
#> 158  58 56.69739
#> 159  76 56.86713
#> 160  41 53.26785
#> 161  16 64.83631
#> 162  96 50.35937
#> 163  26 55.34347
#> 164   4 52.19367
#> 165  24 56.64292
#> 166  25 59.21330
#> 167  57 64.17784
#> 168  68 65.65392
#> 169  51 67.68706
#> 170  13 52.10234
#> 171  66 71.11767
#> 172  64 50.67093
#> 173  27 50.08495
#> 174  25 59.21330
#> 175  37 53.68152
#> 176  32 59.59665
#> 177  79 63.15409
#> 178  24 56.64292
#> 179  48 58.75732
#> 180  85 62.87178
#> 181  27 50.08495
#> 182  48 58.75732
#> 183  85 62.87178
#> 184  35 57.82551
#> 185  52 49.14877
#> 186  79 63.15409
#> 187  19 49.96088
#> 188   1 56.87708
#> 189  97 63.32195
#> 190  53 65.13502
#> 191  14 62.93858
#> 192  58 56.69739
#> 193   5 61.62385
#> 194  97 63.32195
#> 195   3 66.88040
#> 196  59 67.82035
#> 197  76 56.86713
#> 198  20 62.70487
#> 199  31 60.01303
#> 200  46 62.99256