Replicate profiles data.frame

Replicate profiles data.frame

Usage

replicate_profiles(
  profiles,
  replications,
  probabilities = NULL,
  replacement = TRUE,
  seed_used = NULL
)

Arguments

profiles

data.frame of profiles

replications

integer, final number of observations

probabilities

vector of probabilities with the same length as the number of rows of profiles. Does not need to add up to 1 (are reweighted)

replacement

Boolean whether replacement is used

seed_used

Integer with the seed to be used for consistent results

Value

Resampled data.frame of profiles

Examples

replicate_profiles(profiles=data.frame(id=1:100,age=rnorm(100,60,5)),
replications=200,probabilities=rep(1,100))
#>      id      age
#> 1    98 63.41857
#> 2    93 57.73673
#> 3    62 58.54696
#> 4    33 52.36447
#> 5    57 56.60912
#> 6    95 57.45005
#> 7    38 59.91022
#> 8    35 56.27707
#> 9    78 64.10663
#> 10   32 63.68250
#> 11   64 64.76886
#> 12   82 59.34719
#> 13   56 55.69433
#> 14   65 67.35669
#> 15   20 62.33131
#> 16   22 57.17118
#> 17   94 52.99292
#> 18   43 64.88381
#> 19   15 66.13487
#> 20    8 62.92114
#> 21   90 59.60060
#> 22   61 53.41580
#> 23   74 62.59406
#> 24   53 62.99027
#> 25   31 64.99399
#> 26   65 67.35669
#> 27   90 59.60060
#> 28   96 67.81453
#> 29   72 66.64728
#> 30   94 52.99292
#> 31    3 55.80845
#> 32   19 57.55773
#> 33   23 60.63809
#> 34   86 55.65820
#> 35   26 64.57303
#> 36   53 62.99027
#> 37   88 52.77828
#> 38    6 73.08648
#> 39   48 66.48031
#> 40   18 58.75195
#> 41   93 57.73673
#> 42   42 62.77092
#> 43   66 63.91383
#> 44   18 58.75195
#> 45   20 62.33131
#> 46    1 64.01737
#> 47   66 63.91383
#> 48   93 57.73673
#> 49   81 54.85365
#> 50   93 57.73673
#> 51   41 68.50108
#> 52   37 56.27180
#> 53   63 60.24497
#> 54   21 61.11677
#> 55    9 53.80564
#> 56   32 63.68250
#> 57   19 57.55773
#> 58   90 59.60060
#> 59   51 55.04024
#> 60   71 58.11099
#> 61   89 61.27729
#> 62    3 55.80845
#> 63   40 50.20954
#> 64   53 62.99027
#> 65    6 73.08648
#> 66   82 59.34719
#> 67   65 67.35669
#> 68   79 55.95142
#> 69   52 59.60545
#> 70   73 56.14321
#> 71   11 52.44987
#> 72   12 57.07409
#> 73   24 61.84015
#> 74   43 64.88381
#> 75   36 60.77794
#> 76   68 62.96025
#> 77   46 62.05018
#> 78   32 63.68250
#> 79   28 61.88521
#> 80   29 54.25909
#> 81   11 52.44987
#> 82    1 64.01737
#> 83   12 57.07409
#> 84   45 58.00548
#> 85   70 58.13811
#> 86   34 53.24281
#> 87   41 68.50108
#> 88   71 58.11099
#> 89   26 64.57303
#> 90    4 61.97773
#> 91    9 53.80564
#> 92    4 61.97773
#> 93   22 57.17118
#> 94   71 58.11099
#> 95  100 61.55855
#> 96   80 70.29723
#> 97   62 58.54696
#> 98   22 57.17118
#> 99   78 64.10663
#> 100  83 69.36055
#> 101  78 64.10663
#> 102  84 56.34991
#> 103   2 52.41340
#> 104  10 55.54950
#> 105  93 57.73673
#> 106  65 67.35669
#> 107  43 64.88381
#> 108  72 66.64728
#> 109  18 58.75195
#> 110  39 59.89174
#> 111  99 64.12740
#> 112  42 62.77092
#> 113  74 62.59406
#> 114  40 50.20954
#> 115  40 50.20954
#> 116  14 57.11952
#> 117  30 64.70850
#> 118  15 66.13487
#> 119  49 64.33001
#> 120 100 61.55855
#> 121  18 58.75195
#> 122  20 62.33131
#> 123  89 61.27729
#> 124  69 65.64878
#> 125  52 59.60545
#> 126  50 56.75297
#> 127  98 63.41857
#> 128  26 64.57303
#> 129  15 66.13487
#> 130   8 62.92114
#> 131  42 62.77092
#> 132  36 60.77794
#> 133  10 55.54950
#> 134   6 73.08648
#> 135  28 61.88521
#> 136  50 56.75297
#> 137  62 58.54696
#> 138  23 60.63809
#> 139  10 55.54950
#> 140  19 57.55773
#> 141  81 54.85365
#> 142  54 61.51344
#> 143  48 66.48031
#> 144  91 63.91686
#> 145  76 62.90863
#> 146  78 64.10663
#> 147  86 55.65820
#> 148  72 66.64728
#> 149  96 67.81453
#> 150  16 62.30622
#> 151  82 59.34719
#> 152  31 64.99399
#> 153  40 50.20954
#> 154  25 63.04468
#> 155  36 60.77794
#> 156  44 58.49004
#> 157  61 53.41580
#> 158  58 58.82680
#> 159  48 66.48031
#> 160  25 63.04468
#> 161  76 62.90863
#> 162  31 64.99399
#> 163  10 55.54950
#> 164  29 54.25909
#> 165  95 57.45005
#> 166  90 59.60060
#> 167  49 64.33001
#> 168  59 63.38047
#> 169  23 60.63809
#> 170  61 53.41580
#> 171   4 61.97773
#> 172  87 65.96408
#> 173  55 63.65681
#> 174  11 52.44987
#> 175  71 58.11099
#> 176  66 63.91383
#> 177  94 52.99292
#> 178  30 64.70850
#> 179  67 58.18579
#> 180  63 60.24497
#> 181  66 63.91383
#> 182  60 59.66574
#> 183  97 58.26333
#> 184  18 58.75195
#> 185  26 64.57303
#> 186  54 61.51344
#> 187  15 66.13487
#> 188  83 69.36055
#> 189  32 63.68250
#> 190  36 60.77794
#> 191  34 53.24281
#> 192  84 56.34991
#> 193  18 58.75195
#> 194  29 54.25909
#> 195  38 59.91022
#> 196   8 62.92114
#> 197  60 59.66574
#> 198  10 55.54950
#> 199  29 54.25909
#> 200  10 55.54950