Replicate profiles data.frame

Replicate profiles data.frame

Usage

replicate_profiles(
  profiles,
  replications,
  probabilities = NULL,
  replacement = TRUE,
  seed_used = NULL
)

Arguments

profiles

data.frame of profiles

replications

integer, final number of observations

probabilities

vector of probabilities with the same length as the number of rows of profiles. Does not need to add up to 1 (are reweighted)

replacement

Boolean whether replacement is used

seed_used

Integer with the seed to be used for consistent results

Value

Resampled data.frame of profiles

Examples

replicate_profiles(profiles=data.frame(id=1:100,age=rnorm(100,60,5)),
replications=200,probabilities=rep(1,100))
#>      id      age
#> 1     8 64.83631
#> 2    73 62.26069
#> 3    29 53.68152
#> 4    34 59.03215
#> 5    40 58.75732
#> 6    35 59.98832
#> 7    11 49.96088
#> 8    60 65.65392
#> 9    77 62.87178
#> 10   53 64.31948
#> 11   29 53.68152
#> 12   46 58.75759
#> 13   91 53.22700
#> 14   94 61.25549
#> 15   54 57.44199
#> 16   78 58.84650
#> 17   78 58.84650
#> 18   85 57.75288
#> 19   96 64.22396
#> 20   44 49.14877
#> 21   62 56.72692
#> 22   24 59.59665
#> 23   87 57.13262
#> 24   94 61.25549
#> 25   60 65.65392
#> 26   67 63.73726
#> 27    8 64.83631
#> 28   26 60.26786
#> 29   52 51.88512
#> 30   53 64.31948
#> 31   78 58.84650
#> 32   28 51.31351
#> 33   70 55.53916
#> 34   57 61.22590
#> 35   45 65.13502
#> 36   88 50.35937
#> 37   13 59.63312
#> 38   33 53.26785
#> 39   76 62.32756
#> 40   56 50.67093
#> 41   19 50.08495
#> 42   70 55.53916
#> 43   12 62.70487
#> 44   30 62.03154
#> 45   74 61.84000
#> 46    6 62.93858
#> 47   12 62.70487
#> 48   43 67.68706
#> 49   61 64.19335
#> 50   21 65.22638
#> 51   52 51.88512
#> 52   84 55.43344
#> 53   11 49.96088
#> 54   49 64.17784
#> 55   90 51.98730
#> 56   45 65.13502
#> 57   29 53.68152
#> 58    7 60.44821
#> 59   90 51.98730
#> 60   16 56.64292
#> 61   33 53.26785
#> 62   70 55.53916
#> 63   30 62.03154
#> 64   44 49.14877
#> 65   50 56.69739
#> 66   26 60.26786
#> 67   38 62.99256
#> 68    8 64.83631
#> 69   38 62.99256
#> 70  100 66.51682
#> 71   21 65.22638
#> 72   89 63.32195
#> 73   13 59.63312
#> 74   21 65.22638
#> 75   75 58.64806
#> 76   29 53.68152
#> 77   23 60.01303
#> 78   60 65.65392
#> 79   10 52.15650
#> 80   49 64.17784
#> 81   32 65.24229
#> 82   91 53.22700
#> 83   23 60.01303
#> 84   54 57.44199
#> 85    3 52.37253
#> 86   54 57.44199
#> 87   40 58.75732
#> 88    2 67.46406
#> 89   58 71.11767
#> 90   46 58.75759
#> 91   63 64.76980
#> 92    7 60.44821
#> 93   67 63.73726
#> 94   34 59.03215
#> 95   17 59.21330
#> 96   23 60.01303
#> 97   39 59.36894
#> 98   69 61.97612
#> 99   17 59.21330
#> 100  42 57.83179
#> 101  28 51.31351
#> 102  36 59.93585
#> 103  13 59.63312
#> 104  31 52.70173
#> 105  15 58.44466
#> 106  47 62.11160
#> 107  75 58.64806
#> 108   7 60.44821
#> 109  41 60.80164
#> 110  86 64.01466
#> 111  41 60.80164
#> 112  47 62.11160
#> 113  97 59.79014
#> 114  79 65.86121
#> 115  49 64.17784
#> 116 100 66.51682
#> 117  31 52.70173
#> 118  20 58.90200
#> 119  79 65.86121
#> 120  89 63.32195
#> 121  33 53.26785
#> 122  93 64.15619
#> 123  93 64.15619
#> 124  82 56.11315
#> 125  37 60.75974
#> 126  69 61.97612
#> 127  28 51.31351
#> 128  97 59.79014
#> 129  14 57.14491
#> 130  34 59.03215
#> 131  64 61.76476
#> 132  26 60.26786
#> 133  94 61.25549
#> 134  61 64.19335
#> 135  38 62.99256
#> 136  62 56.72692
#> 137  62 56.72692
#> 138  90 51.98730
#> 139  65 61.03300
#> 140  88 50.35937
#> 141  94 61.25549
#> 142  58 71.11767
#> 143  76 62.32756
#> 144  41 60.80164
#> 145  16 56.64292
#> 146  96 64.22396
#> 147  26 60.26786
#> 148   4 64.55359
#> 149  24 59.59665
#> 150  25 64.81491
#> 151  57 61.22590
#> 152  68 56.86713
#> 153  51 67.82035
#> 154  13 59.63312
#> 155  66 65.00557
#> 156  64 61.76476
#> 157  27 57.82551
#> 158  25 64.81491
#> 159  37 60.75974
#> 160  32 65.24229
#> 161  79 65.86121
#> 162  24 59.59665
#> 163  48 64.93827
#> 164  85 57.75288
#> 165  27 57.82551
#> 166  48 64.93827
#> 167  85 57.75288
#> 168  35 59.98832
#> 169  52 51.88512
#> 170  79 65.86121
#> 171  19 50.08495
#> 172   1 61.50527
#> 173  97 59.79014
#> 174  53 64.31948
#> 175  14 57.14491
#> 176  58 71.11767
#> 177   5 52.10234
#> 178  97 59.79014
#> 179   3 52.37253
#> 180  59 61.36688
#> 181  76 62.32756
#> 182  20 58.90200
#> 183  31 52.70173
#> 184  46 58.75759
#> 185  41 60.80164
#> 186  67 63.73726
#> 187  63 64.76980
#> 188  74 61.84000
#> 189  34 59.03215
#> 190  16 56.64292
#> 191   7 60.44821
#> 192  97 59.79014
#> 193   4 64.55359
#> 194  94 61.25549
#> 195  38 62.99256
#> 196  71 63.15409
#> 197  58 71.11767
#> 198  30 62.03154
#> 199   3 52.37253
#> 200  44 49.14877