Replicate profiles data.frame

Replicate profiles data.frame

Usage

replicate_profiles(
  profiles,
  replications,
  probabilities = NULL,
  replacement = TRUE,
  seed_used = NULL
)

Arguments

profiles

data.frame of profiles

replications

integer, final number of observations

probabilities

vector of probabilities with the same length as the number of rows of profiles. Does not need to add up to 1 (are reweighted)

replacement

Boolean whether replacement is used

seed_used

Integer with the seed to be used for consistent results

Value

Resampled data.frame of profiles

Examples

replicate_profiles(profiles=data.frame(id=1:100,age=rnorm(100,60,5)),
replications=200,probabilities=rep(1,100))
#>      id      age
#> 1    73 61.67248
#> 2    11 58.98776
#> 3    12 68.42248
#> 4    24 60.77607
#> 5    43 59.38907
#> 6    36 64.33908
#> 7    68 57.54275
#> 8    46 61.27541
#> 9    32 52.17178
#> 10   28 56.41231
#> 11   29 61.80847
#> 12   11 58.98776
#> 13    1 51.79709
#> 14   12 68.42248
#> 15   45 55.19362
#> 16   70 57.91901
#> 17   34 62.57041
#> 18   41 64.15074
#> 19   71 57.39728
#> 20   26 52.33467
#> 21    4 62.96200
#> 22    9 57.35178
#> 23    4 62.96200
#> 24   22 56.01136
#> 25   71 57.39728
#> 26  100 63.71614
#> 27   80 68.04312
#> 28   62 65.42464
#> 29   22 56.01136
#> 30   78 60.15535
#> 31   83 65.07853
#> 32   78 60.15535
#> 33   84 60.18867
#> 34    2 62.53961
#> 35   10 56.59363
#> 36   93 55.65531
#> 37   65 70.84672
#> 38   43 59.38907
#> 39   72 64.25292
#> 40   18 60.57015
#> 41   39 51.07818
#> 42   99 64.26364
#> 43   42 63.03673
#> 44   74 55.85324
#> 45   40 54.81427
#> 46   40 54.81427
#> 47   14 59.22012
#> 48   30 66.99502
#> 49   15 59.76800
#> 50   49 57.31175
#> 51  100 63.71614
#> 52   18 60.57015
#> 53   20 55.40334
#> 54   89 73.46186
#> 55   69 67.67458
#> 56   52 56.91888
#> 57   50 57.73787
#> 58   98 60.17040
#> 59   26 52.33467
#> 60   15 59.76800
#> 61    8 54.28217
#> 62   42 63.03673
#> 63   36 64.33908
#> 64   10 56.59363
#> 65    6 60.60810
#> 66   28 56.41231
#> 67   50 57.73787
#> 68   62 65.42464
#> 69   23 63.34111
#> 70   10 56.59363
#> 71   19 60.31959
#> 72   81 67.14927
#> 73   54 62.41299
#> 74   48 64.65180
#> 75   91 66.80038
#> 76   76 52.27458
#> 77   78 60.15535
#> 78   86 55.14963
#> 79   72 64.25292
#> 80   96 62.53479
#> 81   16 55.23186
#> 82   82 55.25830
#> 83   31 61.86349
#> 84   40 54.81427
#> 85   25 60.64344
#> 86   36 64.33908
#> 87   44 64.66563
#> 88   61 63.16086
#> 89   58 67.02039
#> 90   48 64.65180
#> 91   25 60.64344
#> 92   76 52.27458
#> 93   31 61.86349
#> 94   10 56.59363
#> 95   29 61.80847
#> 96   95 66.07354
#> 97   90 66.96197
#> 98   49 57.31175
#> 99   59 66.88559
#> 100  23 63.34111
#> 101  61 63.16086
#> 102   4 62.96200
#> 103  87 58.85486
#> 104  55 62.71072
#> 105  11 58.98776
#> 106  71 57.39728
#> 107  66 60.69565
#> 108  94 57.46722
#> 109  30 66.99502
#> 110  67 66.88163
#> 111  63 66.78230
#> 112  66 60.69565
#> 113  60 65.30198
#> 114  97 49.52514
#> 115  18 60.57015
#> 116  26 52.33467
#> 117  54 62.41299
#> 118  15 59.76800
#> 119  83 65.07853
#> 120  32 52.17178
#> 121  36 64.33908
#> 122  34 62.57041
#> 123  84 60.18867
#> 124  18 60.57015
#> 125  29 61.80847
#> 126  38 59.18660
#> 127   8 54.28217
#> 128  60 65.30198
#> 129  10 56.59363
#> 130  29 61.80847
#> 131  10 56.59363
#> 132  93 55.65531
#> 133  25 60.64344
#> 134  97 49.52514
#> 135  33 59.74153
#> 136  45 55.19362
#> 137  93 55.65531
#> 138  86 55.14963
#> 139  97 49.52514
#> 140  33 59.74153
#> 141  90 66.96197
#> 142  31 61.86349
#> 143  46 61.27541
#> 144  46 61.27541
#> 145  44 64.66563
#> 146  88 55.12577
#> 147  80 68.04312
#> 148  38 59.18660
#> 149  35 62.74950
#> 150  98 60.17040
#> 151  90 66.96197
#> 152  45 55.19362
#> 153  91 66.80038
#> 154  21 64.50668
#> 155  77 61.16615
#> 156  20 55.40334
#> 157  50 57.73787
#> 158  15 59.76800
#> 159  64 61.81212
#> 160  36 64.33908
#> 161  61 63.16086
#> 162  61 63.16086
#> 163  10 56.59363
#> 164  76 52.27458
#> 165  74 55.85324
#> 166  82 55.25830
#> 167  52 56.91888
#> 168  28 56.41231
#> 169  55 62.71072
#> 170  23 63.34111
#> 171  27 61.01180
#> 172  36 64.33908
#> 173  33 59.74153
#> 174  92 58.17292
#> 175  52 56.91888
#> 176  90 66.96197
#> 177  21 64.50668
#> 178 100 63.71614
#> 179  93 55.65531
#> 180  77 61.16615
#> 181  90 66.96197
#> 182  29 61.80847
#> 183  94 57.46722
#> 184  86 55.14963
#> 185  48 64.65180
#> 186  71 57.39728
#> 187  28 56.41231
#> 188  21 64.50668
#> 189  43 59.38907
#> 190  25 60.64344
#> 191  11 58.98776
#> 192  29 61.80847
#> 193   2 62.53961
#> 194  64 61.81212
#> 195   9 57.35178
#> 196  37 63.42180
#> 197  62 65.42464
#> 198  98 60.17040
#> 199  63 66.78230
#> 200  54 62.41299