Replicate profiles data.frame

Replicate profiles data.frame

Usage

replicate_profiles(
  profiles,
  replications,
  probabilities = NULL,
  replacement = TRUE,
  seed_used = NULL
)

Arguments

profiles

data.frame of profiles

replications

integer, final number of observations

probabilities

vector of probabilities with the same length as the number of rows of profiles. Does not need to add up to 1 (are reweighted)

replacement

Boolean whether replacement is used

seed_used

Integer with the seed to be used for consistent results

Value

Resampled data.frame of profiles

Examples

replicate_profiles(profiles=data.frame(id=1:100,age=rnorm(100,60,5)),
replications=200,probabilities=rep(1,100))
#>      id      age
#> 1    50 58.75759
#> 2     9 52.10234
#> 3    15 49.96088
#> 4    91 57.13262
#> 5    73 61.97612
#> 6    89 57.75288
#> 7    92 50.35937
#> 8    49 65.13502
#> 9     8 64.55359
#> 10   73 61.97612
#> 11   29 64.81491
#> 12   34 62.03154
#> 13   40 59.93585
#> 14   35 52.70173
#> 15   11 60.44821
#> 16   60 50.67093
#> 17   77 62.26069
#> 18   53 64.17784
#> 19   29 64.81491
#> 20   46 57.83179
#> 21   91 57.13262
#> 22   94 51.98730
#> 23   54 56.69739
#> 24   78 61.84000
#> 25   78 61.84000
#> 26   85 56.69004
#> 27   96 44.91034
#> 28   44 58.75732
#> 29   62 71.11767
#> 30   24 58.90200
#> 31   87 62.56769
#> 32   94 51.98730
#> 33   60 50.67093
#> 34   67 64.76980
#> 35    8 64.55359
#> 36   26 69.38665
#> 37   52 64.93827
#> 38   53 64.17784
#> 39   78 61.84000
#> 40   28 59.59665
#> 41   70 65.00557
#> 42   57 64.31948
#> 43   45 60.80164
#> 44   88 55.43344
#> 45   13 60.39406
#> 46   33 53.68152
#> 47   76 57.83647
#> 48   56 51.88512
#> 49   19 58.44466
#> 50   70 65.00557
#> 51   12 64.83631
#> 52   30 60.26786
#> 53   74 55.53916
#> 54    6 67.46406
#> 55   12 64.83631
#> 56   43 59.36894
#> 57   61 61.22590
#> 58   21 59.21330
#> 59   52 64.93827
#> 60   84 66.96351
#> 61   11 60.44821
#> 62   49 65.13502
#> 63   90 64.01466
#> 64   45 60.80164
#> 65   29 64.81491
#> 66    7 52.37253
#> 67   90 64.01466
#> 68   16 62.70487
#> 69   33 53.68152
#> 70   70 65.00557
#> 71   30 60.26786
#> 72   44 58.75732
#> 73   50 58.75759
#> 74   26 69.38665
#> 75   38 59.03215
#> 76    8 64.55359
#> 77   38 59.03215
#> 78  100 64.22396
#> 79   21 59.21330
#> 80   89 57.75288
#> 81   13 60.39406
#> 82   21 59.21330
#> 83   75 63.15409
#> 84   29 64.81491
#> 85   23 50.08495
#> 86   60 50.67093
#> 87   10 62.93858
#> 88   49 65.13502
#> 89   32 51.31351
#> 90   91 57.13262
#> 91   23 50.08495
#> 92   54 56.69739
#> 93    3 64.38892
#> 94   54 56.69739
#> 95   40 59.93585
#> 96    2 59.21605
#> 97   58 57.44199
#> 98   46 57.83179
#> 99   63 61.36688
#> 100   7 52.37253
#> 101  67 64.76980
#> 102  34 62.03154
#> 103  17 59.63312
#> 104  23 50.08495
#> 105  39 59.98832
#> 106  69 61.03300
#> 107  17 59.63312
#> 108  42 62.99256
#> 109  28 59.59665
#> 110  36 65.24229
#> 111  13 60.39406
#> 112  31 57.82551
#> 113  15 49.96088
#> 114  47 67.68706
#> 115  75 63.15409
#> 116   7 52.37253
#> 117  41 60.75974
#> 118  86 56.11315
#> 119  41 60.75974
#> 120  47 67.68706
#> 121  97 64.15619
#> 122  79 58.64806
#> 123  49 65.13502
#> 124 100 64.22396
#> 125  31 57.82551
#> 126  20 56.64292
#> 127  79 58.64806
#> 128  89 57.75288
#> 129  33 53.68152
#> 130  93 63.32195
#> 131  93 63.32195
#> 132  82 58.84650
#> 133  37 53.26785
#> 134  69 61.03300
#> 135  28 59.59665
#> 136  97 64.15619
#> 137  14 52.15650
#> 138  34 62.03154
#> 139  64 65.65392
#> 140  26 69.38665
#> 141  94 51.98730
#> 142  61 61.22590
#> 143  38 59.03215
#> 144  62 71.11767
#> 145  62 71.11767
#> 146  90 64.01466
#> 147  65 64.19335
#> 148  88 55.43344
#> 149  94 51.98730
#> 150  58 57.44199
#> 151  76 57.83647
#> 152  41 60.75974
#> 153  16 62.70487
#> 154  96 44.91034
#> 155  26 69.38665
#> 156   4 63.75083
#> 157  24 58.90200
#> 158  25 65.22638
#> 159  57 64.31948
#> 160  68 61.76476
#> 161  51 62.11160
#> 162  13 60.39406
#> 163  66 56.72692
#> 164  64 65.65392
#> 165  27 60.01303
#> 166  25 65.22638
#> 167  37 53.26785
#> 168  32 51.31351
#> 169  79 58.64806
#> 170  24 58.90200
#> 171  48 49.14877
#> 172  85 56.69004
#> 173  27 60.01303
#> 174  48 49.14877
#> 175  85 56.69004
#> 176  35 52.70173
#> 177  52 64.93827
#> 178  79 58.64806
#> 179  19 58.44466
#> 180   1 61.62385
#> 181  97 64.15619
#> 182  53 64.17784
#> 183  14 52.15650
#> 184  58 57.44199
#> 185   5 61.50527
#> 186  97 64.15619
#> 187   3 64.38892
#> 188  59 50.41317
#> 189  76 57.83647
#> 190  20 56.64292
#> 191  31 57.82551
#> 192  46 57.83179
#> 193  41 60.75974
#> 194  67 64.76980
#> 195  63 61.36688
#> 196  74 55.53916
#> 197  34 62.03154
#> 198  16 62.70487
#> 199   7 52.37253
#> 200  97 64.15619